Update ddl_script_bronze_layer
This is the code which I have used in Microsoft SQL Server Management Studio 2022 for creating the Bronze Layer and the tables in it with the csv files which were provided by the client. The tables will be named after the csv files and the columns inside the tables will be named as they are listed in the csv files.