Update procedure_load_data_bronze_layer
This file contains the code for loading the data from the csv files on to the tables which I created in the bronze layer. I added few fancy codes so that when the data is loaded the message window shows extra things like "Loading Bronze Layer" "Inserting Data in" 
This makes the result look good that what type of data was loaded in which table.